Move Cache to public package in persistent cache
