+++ subprojects/language-java/src/main/java/org/gradle/api/internal/tasks/compile/incremental/analyzer/ClassAnalysisCache.java
@@ -17,7 +17,7 @@
 package org.gradle.api.internal.tasks.compile.incremental.analyzer;
 
 import org.gradle.api.internal.tasks.compile.incremental.deps.ClassAnalysis;
-import org.gradle.cache.internal.Cache;
+import org.gradle.cache.Cache;
 import org.gradle.internal.hash.HashCode;
 
 public interface ClassAnalysisCache extends Cache<HashCode, ClassAnalysis> {
