+++ subprojects/persistent-cache/src/main/java/org/gradle/cache/internal/CacheAccessSerializer.java
@@ -15,6 +15,7 @@
  */
 package org.gradle.cache.internal;
 
+import org.gradle.cache.Cache;
 import org.gradle.internal.concurrent.Synchronizer;
 
 import java.util.function.Function;
