+++ subprojects/persistent-cache/src/main/java/org/gradle/cache/internal/MinimalPersistentCache.java
@@ -16,6 +16,7 @@
 
 package org.gradle.cache.internal;
 
+import org.gradle.cache.Cache;
 import org.gradle.cache.PersistentIndexedCache;
 
 import java.util.function.Function;
